test: exercise symbolic link following
This patch adds an integration test that ensures symbolic links are
followed.

Refs: #22
