+++ .gitignore
@@ -1,3 +1,4 @@
 /target
+test_helpers/target
 **/*.rs.bk
 Cargo.lock
